[diffusion] chore: clean srt imports
