+++ python/sglang/multimodal_gen/runtime/entrypoints/utils.py
@@ -103,3 +103,5 @@ def post_process_sample(
             logger.info(f"No output path provided, output not saved")
 
     return frames
+
+
