+++ python/sglang/multimodal_gen/runtime/loader/component_loader.py
@@ -149,7 +149,7 @@ class ComponentLoader(ABC):
         """
         gpu_mem_before_loading = current_platform.get_available_gpu_memory()
         logger.info(
-            "Loading %s from %s. avail mem: %.2f GB",
+            "Loading %s. avail mem: %.2f GB",
             module_name,
             component_model_path,
             gpu_mem_before_loading,
